Move navigation related stuff into feature folder
