+++ app/src/lib/navigation/BranchItem.svelte
