+++ app/src/lib/navigation/Footer.svelte
@@ -1,5 +1,5 @@
 <script lang="ts">
-	import Button from './Button.svelte';
+	import Button from '../components/Button.svelte';
 	import AccountLink from '$lib/components/AccountLink.svelte';
 	import * as events from '$lib/utils/events';
 	import { goto } from '$app/navigation';
