+++ app/src/lib/navigation/ProjectAvatar.svelte
