+++ app/src/lib/navigation/ProjectSelector.svelte
